refactor to make two icons into separate components, make isInBatch13 and isCheckedIn into variables that hold the booleans, refactor state machine into a useMemo and switch statement into an object mapping of type Record<string, JSX.element>
